TG-52: refactor forms in one
