+++ src/routes/admins.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Admins from 'Components/Admins';
-import Form from 'Components/Share/Form';
+// import Form from 'Components/Share/Form';
 import CreateForm from 'Components/Share/CreateForm';
 
 const AdminsRoutes = () => {
@@ -8,7 +8,7 @@ const AdminsRoutes = () => {
     <Router>
       <Switch>
         <Route exact path="/admins" component={Admins} />
-        <Route path="/admins/form/:id" render={() => <Form />} />
+        <Route path="/admins/form/:id" render={() => <CreateForm />} />
         <Route path="/admins/new" render={() => <CreateForm />} />
       </Switch>
     </Router>
