+++ src/routes/employees.jsx
@@ -1,6 +1,6 @@
 import { BrowserRouter as Router, Switch, Route } from 'react-router-dom';
 import Employees from 'Components/Employees';
-import Form from 'Components/Share/Form';
+// import Form from 'Components/Share/Form';
 import CreateForm from 'Components/Share/CreateForm';
 
 const EmployeesRoutes = () => {
@@ -8,7 +8,7 @@ const EmployeesRoutes = () => {
     <Router>
       <Switch>
         <Route exact path="/employees" component={Employees} />
-        <Route path="/employees/form/:id" render={() => <Form />} />
+        <Route path="/employees/form/:id" render={() => <CreateForm />} />
         <Route path="/employees/new" render={() => <CreateForm />} />
       </Switch>
     </Router>
